Move mojos to mojo package
