+++ src/main/java/io/brooklyn/maven/mojo/AbstractInvokeBrooklynMojo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.brooklyn.maven;
+package io.brooklyn.maven.mojo;
 
 import static com.google.common.base.Preconditions.checkNotNull;
