+++ src/main/java/io/brooklyn/maven/mojo/StopApplicationMojo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.brooklyn.maven;
+package io.brooklyn.maven.mojo;
 
 import java.net.URL;
