+++ src/test/java/io/brooklyn/maven/mojo/StopBrooklynMojoTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.brooklyn.maven;
+package io.brooklyn.maven.mojo;
 
 import static org.junit.Assert.assertEquals;
 
@@ -23,6 +23,8 @@ import org.junit.Test;
 
 import com.squareup.okhttp.mockwebserver.MockResponse;
 import com.squareup.okhttp.mockwebserver.RecordedRequest;
+
+import io.brooklyn.maven.AbstractBrooklynMojoTest;
 
 public class StopBrooklynMojoTest extends AbstractBrooklynMojoTest {
