+++ src/test/resources/io/brooklyn/maven/mojo/failed-task-status.json
